[misc] resolve implicit float to int cast warnings in PawfectPHPCommand

+++ src/PawfectPHPCommand.php
@@ -225,8 +225,8 @@ class PawfectPHPCommand extends Command
             $analysis->getPassCount(),
             $analysis->getFailCount(),
             $analysis->getExceptionCount(),
-            $analysis->getWarnCount(),
-            sprintf('%02d:%02d:%02d', ($duration / 3600), ($duration / 60 % 60), $duration % 60)
+            $analysis->getWarnCount(), 
+            sprintf('%02d:%02d:%02d', (int)($duration / 3600), ((int)($duration / 60) % 60), $duration % 60)
         ));
 
         if ($analysis->getFailCount() > 0) {
